feat: modularises pubsub
License: MIT
